requirements.txt
Package                   Version
------------------------- -------------
aniso8601                 9.0.1
attrs                     23.1.0
beautifulsoup4            4.10.0
blinker                   1.6.3
certifi                   2023.7.22
chardet                   4.0.0
charset-normalizer        3.3.0
click                     8.1.7
command-not-found         0.3
cryptography              3.4.8
dbus-python               1.2.18
distlib                   0.3.7
distro                    1.7.0
distro-info               1.1build1
filelock                  3.13.1
flasgger                  0.9.7.1
Flask                     2.3.3
Flask-Cors                4.0.0
Flask-JWT-Extended        4.5.3
flask-marshmallow         0.15.0
flask-restx               1.2.0
Flask-SQLAlchemy          3.1.1
greenlet                  3.0.0
gyp                       0.1
html5lib                  1.1
httplib2                  0.20.2
idna                      3.4
importlib-metadata        4.6.4
importlib-resources       6.1.1
itsdangerous              2.1.2
jeepney                   0.7.1
Jinja2                    3.1.2
jsonschema                4.17.3
jsonschema-specifications 2023.11.1
keyring                   23.5.0
launchpadlib              1.10.16
lazr.restfulclient        0.14.4
lazr.uri                  1.0.6
lxml                      4.8.0
MarkupSafe                2.1.3
marshmallow               3.20.1
marshmallow-sqlalchemy    0.29.0
mistune                   3.0.2
more-itertools            8.10.0
mysqlclient               2.1.1
netifaces                 0.11.0
oauthlib                  3.2.0
packaging                 23.2
passlib                   1.7.4
pip                       22.0.2
pipenv                    2023.11.15
platformdirs              3.11.0
pycodestyle               2.8.0
PyGObject                 3.42.1
PyJWT                     2.3.0
pyparsing                 2.4.7
pyrsistent                0.20.0
python-apt                2.4.0+ubuntu1
python-decouple           3.8
python-dotenv             1.0.0
pytz                      2023.3.post1
PyYAML                    5.4.1
referencing               0.31.0
requests                  2.31.0
rpds-py                   0.13.0
SecretStorage             3.3.1
setuptools                68.2.2
six                       1.16.0
soupsieve                 2.3.1
SQLAlchemy                2.0.22
systemd-python            234
typing_extensions         4.8.0
ubuntu-advantage-tools    8001
ufw                       0.36.1
unattended-upgrades       0.1
urllib3                   2.0.7
virtualenv                20.24.6
wadllib                   1.3.6
webencodings              0.5.1
Werkzeug                  2.3.8
wheel                     0.37.1
zipp                      1.0.0
